Move handler code into src sub-folder.
