+++ registration/api/src/schema.json
